在user_service/utils提供了统一查userInfo的方法
