+++ internal/service/relation_service/query_follower_list.go
@@ -54,7 +54,7 @@ func (f *QueryFollowerInfoFlow) prepareFollowerInfo() error {
 	if err != nil {
 		return err
 	}
-	userInfoList, err := user_service.QueryUserList(f.hostId, &idList)
+	userInfoList, err := user_service.QueryUserInfoList(f.hostId, &idList)
 	if err != nil {
 		return errors.New("DB Find Follower Error")
 	}
